ai/requirements.txt
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0

# ASR - Automatic Speech Recognition
faster-whisper>=1.0.0